correct changes

+++ main.go
@@ -1,8 +1,8 @@
 package main
 
 func main() {
-	var balance int32 = 1500000000
-	var transfer int32 = 1000000000
+	var balance int64 = 1500000000
+	var transfer int64 = 1000000000
 	total := balance + transfer
 	println(total)
 }
